Update bladeAdd.sh
Added color and case handling with errors is done outside of the functions themselves. Also added a failure variable to flag any errors and prevent the blade add portion from executing if this flag is set. This should allow for testing this script on inducted blades which will just throw an error but allow the rest of the checks to execute without exiting the program.

Future plans are to change the command line args to match the -n1.1-15 style that fbupgrade uses, add more checks, add a -h option, and also allow for a scan to detect any "unclaimed" blades that are on multiple chassis to expand functionality and automation.